chore: add dish ingredient inside of dish

+++ src/db/migration/3_create_dish_ingredient _table.sql
@@ -2,7 +2,7 @@ create table if not exists "dish_ingredient"
 (
     "id_dish" varchar not null references "dish"("id"),
     "id_ingredient" varchar not null references "ingredient"("id"),
-    "required_quantity" decimal(10,2) check ("required_quantity" > 0) not null,
+    "required_quantity" float check ("required_quantity" > 0) not null,
     "unit" unit not null,
     primary key ("id_dish", "id_ingredient")
 );
